[WIP] Extract AS::AbstractClass from AR::Inheritance
